Crate fasten_inspect

Marks up your reads with useful information like read length

Examples

Quick validation with stderr message

cat file.fastq | fasten_inspect > markedup.fastq
cat file.fastq | fasten_inspect --paired-end > markedup-paired.fastq

The resulting marked-up fastq file will have deflines like

@read0/1 id-at:1 seq-length:100 seq-invalid-chars: id-plus:1 qual-invalid-chars: avg-qual:20.93 qual-length:100 read-pair:1

Usage

fasten_inspect: Marks up your reads with useful information like read length

Usage: fasten_inspect [-h] [-n INT] [-p] [--verbose] [--version]

Options:
   -h, --help          Print this help menu.
   -n, --numcpus INT   Number of CPUs (default: 1)
   -p, --paired-end    The input reads are interleaved paired-end
       --verbose       Print more status messages
       --version       Print the version of Fasten and exit

The fields will be found on the defline of the sequence and include:

key	type	example	note

id-at	boolean (1 or 0)	id-at:1	Whether or not the @ was first character, first line

seq-invalid-chars	string	seq-invalid-chars:$$%

qual-invalid-chars	string	qual-invalid-chars:[<

seq-length	int	seq-length:100

id-plus	boolean	id-plus:1	Whether or not the + was first character, 3rd line

avg-qual	float	avg-qual:17.52

qual-length	int	qual-length:100	Length of the quality score line

Functions

main 🔒

validate_reads 🔒

marks up reads from stdin